+++ components/organism/RankingPage/RankingBoardHeader/RankingBoardHeader.tsx
@@ -1,3 +1,4 @@
+'use client';
 import React from 'react';
 
 export const RankingBoardHeader = () => {
