+++ components/organism/RankingPage/RankingFilters/RankingFilters.tsx
@@ -1,17 +1,17 @@
 'use client';
+
 import React, { ChangeEvent, useCallback } from 'react';
 import TierFilter from '../../CompetitionPage/CompetitionHomePage/TierFilter/TierFilter';
 import DropdownIcon from '@/app/_asset/icons/dropdown.svg';
 import type { Tier } from '@/@types/tier';
-import { UserRanking } from '@/@types/ranking';
 import { usePathname, useRouter, useSearchParams } from 'next/navigation';
 
 const RankingFilters = ({
   tiers,
-  usersRankingData,
+  clubNameArr,
 }: {
   tiers: Tier[];
-  usersRankingData: UserRanking[];
+  clubNameArr: string[];
 }) => {
   const router = useRouter();
   const pathname = usePathname();
@@ -30,17 +30,6 @@ const RankingFilters = ({
   const handleChange = (event: ChangeEvent<HTMLSelectElement>) => {
     router.replace(pathname + '?' + createQueryString('club', event.target.value));
   };
-
-  const clubNameSet = new Set(
-    usersRankingData
-      ? usersRankingData.map((data) => {
-          if (!data.club) return '무소속';
-          return data.club?.name;
-        })
-      : [],
-  );
-
-  const clubNameArr = Array.from(clubNameSet).sort((a, b) => a.localeCompare(b, 'ko-KR'));
 
   return (
     <div className="flex gap-[16px]">
